// src/_shared/enums/transporte-status.enum.ts
export enum TransporteStatus {
  NAO_INICIADO = 'NAO_INICIADO',
  EM_PROGRESSO = 'EM_PROGRESSO',
  CONCLUIDO = 'CONCLUIDO',
  EM_PAUSA = 'EM_PAUSA',
}

export enum StatusTransporte {
  AGUARDANDO_SEPARACAO = 'AGUARDANDO_SEPARACAO',
  EM_SEPARACAO = 'EM_SEPARACAO',
  SEPARACAO_CONCLUIDA = 'SEPARACAO_CONCLUIDA',
  EM_CONFERENCIA = 'EM_CONFERENCIA',
  CONFERENCIA_CONCLUIDA = 'CONFERENCIA_CONCLUIDA',
  EM_CARREGAMENTO = 'EM_CARREGAMENTO',
  CARREGAMENTO_CONCLUIDO = 'CARREGAMENTO_CONCLUIDO',
  FATURADO = 'FATURADO',
  LIBERADO_PORTARIA = 'LIBERADO_PORTARIA',
  CANCELADO = 'CANCELADO',
}